refactor(core): move TT UI components to own directory
