+++ core/src/apps/cardano/layout.py
@@ -10,9 +10,9 @@ from trezor.messages import (
     CardanoPoolOwnerType,
 )
 from trezor.strings import format_amount
-from trezor.ui.button import ButtonDefault
-from trezor.ui.scroll import Paginated
-from trezor.ui.text import Text
+from trezor.ui.components.tt.button import ButtonDefault
+from trezor.ui.components.tt.scroll import Paginated
+from trezor.ui.components.tt.text import Text
 from trezor.utils import chunks
 
 from apps.common.confirm import confirm, require_confirm, require_hold_to_confirm
