+++ core/src/apps/eos/actions/layout.py
@@ -3,8 +3,8 @@ from ubinascii import hexlify
 
 from trezor import ui
 from trezor.messages import ButtonRequestType
-from trezor.ui.scroll import Paginated
-from trezor.ui.text import Text
+from trezor.ui.components.tt.scroll import Paginated
+from trezor.ui.components.tt.text import Text
 from trezor.utils import chunks
 
 from .. import helpers
