+++ core/src/apps/management/recovery_device/keyboard_bip39.py
@@ -1,11 +1,16 @@
 from trezor import io, loop, res, ui, workflow
 from trezor.crypto import bip39
 from trezor.ui import display
-from trezor.ui.button import Button, ButtonClear, ButtonMono, ButtonMonoConfirm
+from trezor.ui.components.tt.button import (
+    Button,
+    ButtonClear,
+    ButtonMono,
+    ButtonMonoConfirm,
+)
 
 if False:
     from typing import Optional, Tuple
-    from trezor.ui.button import ButtonContent, ButtonStyleStateType
+    from trezor.ui.components.tt.button import ButtonContent, ButtonStyleStateType
 
 
 def compute_mask(text: str) -> int:
