+++ core/src/apps/management/reset_device/layout.py
@@ -3,13 +3,12 @@ import ubinascii
 from trezor import ui, utils
 from trezor.crypto import random
 from trezor.messages import BackupType, ButtonRequestType
-from trezor.ui.button import Button, ButtonDefault
-from trezor.ui.checklist import Checklist
-from trezor.ui.info import InfoConfirm
-from trezor.ui.loader import LoadingAnimation
-from trezor.ui.num_input import NumInput
-from trezor.ui.scroll import Paginated
-from trezor.ui.text import Text
+from trezor.ui.components.tt.button import Button, ButtonDefault
+from trezor.ui.components.tt.checklist import Checklist
+from trezor.ui.components.tt.info import InfoConfirm
+from trezor.ui.components.tt.num_input import NumInput
+from trezor.ui.components.tt.scroll import Paginated
+from trezor.ui.components.tt.text import Text
 
 from apps.common.confirm import confirm, require_confirm, require_hold_to_confirm
 from apps.common.layout import show_success
