+++ core/src/apps/tezos/layout.py
@@ -1,8 +1,8 @@
 from trezor import ui
 from trezor.messages import ButtonRequestType
 from trezor.strings import format_amount
-from trezor.ui.scroll import Paginated
-from trezor.ui.text import Text
+from trezor.ui.components.tt.scroll import Paginated
+from trezor.ui.components.tt.text import Text
 from trezor.utils import chunks
 
 from apps.common.confirm import require_confirm, require_hold_to_confirm
