+++ core/src/apps/webauthn/list_resident_credentials.py
@@ -4,7 +4,7 @@ from trezor.messages.WebAuthnCredentials import WebAuthnCredentials
 from trezor.messages.WebAuthnListResidentCredentials import (
     WebAuthnListResidentCredentials,
 )
-from trezor.ui.text import Text
+from trezor.ui.components.tt.text import Text
 
 from apps.common.confirm import require_confirm
