+++ core/src/trezor/ui/components/tt/passphrase.py
@@ -2,12 +2,13 @@ from micropython import const
 
 from trezor import io, loop, res, ui, workflow
 from trezor.ui import display
-from trezor.ui.button import Button, ButtonClear, ButtonConfirm
-from trezor.ui.swipe import SWIPE_HORIZONTAL, SWIPE_LEFT, Swipe
+
+from .button import Button, ButtonClear, ButtonConfirm
+from .swipe import SWIPE_HORIZONTAL, SWIPE_LEFT, Swipe
 
 if False:
     from typing import Iterable, List, Optional, Tuple
-    from trezor.ui.button import ButtonContent, ButtonStyleStateType
+    from .button import ButtonContent, ButtonStyleStateType
 
 SPACE = res.load(ui.ICON_SPACE)
