+++ core/src/trezor/ui/components/tt/swipe.py
